fixed migration and db files

+++ src/payment/models.py
@@ -10,7 +10,7 @@ class Payment(models.Model):
     )
     student_couse = models.ForeignKey(StudentCourse, on_delete=models.CASCADE)
     amount = models.DecimalField(max_digits=10, decimal_places=2,null=False,blank=False)
-    type = models.IntegerField(choices=pay_type,default=0)
+    type = models.SmallIntegerField(choices=pay_type,default=0)
     payment_date = models.DateField(auto_now_add=True)
 
     def __str__(self):
